작성자만 수정/삭제, login_required 추가

+++ reviews/views.py
@@ -2,8 +2,10 @@ from django.shortcuts import render, redirect
 from restaurants.models import Restaurant
 from .models import Review, Comment
 from .forms import ReviewForm, CommentForm
+from django.contrib.auth.decorators import login_required
 
 
+@login_required
 def create(request, restaurant_id):
     restaurant = Restaurant.objects.get(pk=restaurant_id)
     if request.method == 'POST':
@@ -38,15 +40,21 @@ def detail(request, restaurant_id, review_id):
     return render(request, 'reviews/detail.html', context)
     
     
+@login_required
 def delete(request, restaurant_id, review_id):
     review = Review.objects.get(pk=review_id)
+    if request.user != review.user:
+        return redirect('reviews:detail', restaurant_id=restaurant_id, review_id=review_id)
     review.delete()
     return redirect('restaurants:detail', restaurant_id=restaurant_id)
     
     
+@login_required
 def update(request, restaurant_id, review_id):
     review = Review.objects.get(pk=review_id)
     restaurant = Restaurant.objects.get(pk=restaurant_id)
+    if request.user != review.user:
+        return redirect('reviews:detail', restaurant_id=restaurant_id, review_id=review_id)
     if request.method == 'POST':
         review_form = ReviewForm(request.POST, instance=review)
         if review_form.is_valid():
@@ -66,6 +74,7 @@ def update(request, restaurant_id, review_id):
     return render(request, 'reviews/update.html', context)
     
     
+@login_required
 def likes(request, restaurant_id, review_id):
     review = Review.objects.get(pk=review_id)
     if review.like_users.filter(pk=request.user.pk).exists():
@@ -74,6 +83,8 @@ def likes(request, restaurant_id, review_id):
         review.like_users.add(request.user)
     return redirect('reviews:detail', restaurant_id=restaurant_id, review_id=review_id)
 
+
+@login_required
 def comment_create(request, restaurant_id, review_id):
     restaurant = Restaurant.objects.get(pk=restaurant_id)
     review = Review.objects.get(pk=review_id)
@@ -91,7 +102,9 @@ def comment_create(request, restaurant_id, review_id):
     return redirect('reviews:detail', restaurant_id=restaurant_id, review_id=review_id, context=context)
 
 
+@login_required
 def comment_delete(request, restaurant_id, review_id, comment_pk):
     comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
+    if request.user == comment.user:
+        comment.delete()
     return redirect('reviews:detail', restaurant_id=restaurant_id, review_id=review_id)
